Update Backus-Naur Form.bnf
Added spaces between tokens when necessary. these could be removed though if we make the lexical automatically split tokens. just to test the logic on a BNF playground the spaces made the test input look as they should in a real environment. also fixed the for loop to have a proper assignment and made the variables less strict for naming. We can trim the letters and digits down in the EBNF